Add model for world to initialize a grid
The grid should be initialized upon creating a world, simply creating a 2D array to represent rows and columns. The values of the cells will represent the state of that cell, which also influences the UI for the grid. Also include methods to get/set data in cells for later.