Add .scss file for grid-based UI elements to be used in views
The grid utilizes a container that stores multiple rows of cells. Cells are empty or filled based on their data, which will need to be updated upon the creation of seams for the AI team. The grid will be centered, as well.

A toolbar is also defined to hold world information. A back button will also be added to this bar.