Update login/signup for easier user flow. The main action (login on the login page, create account on the signup page) is now the topmost button, and is a darker color. This helps user experience by drawing their attention to the main desired button.
Also, note how I created the 'back to log in' and 'create account' buttons, since they are just links and do not actually submit the form. To do this while keeping the formatting of the button, I had to use a button as a wrapper within the href, instead of trying to create a button itself. The formatting will not directly apply to an href, and buttons as a base would submit the form. This is what I found to be the easiest way to do this.